// src/app/model/udc-name.enum.ts
export enum UdcNameEnum {
    Eisdiele_Titelbild_1920x400 = 'Eisdiele_Titelbild_1920x400',
    Eisdiele_Clubboard_348x180 = 'Eisdiele_Clubboard_348x180',
    Eisdiele_Partnersuche_348x180 = 'Eisdiele_Partnersuche_348x180',
    Eisdiele_Erlebnis_Chats_348x180 = 'Eisdiele_Erlebnis_Chats_348x180',
    Eisdiele_Wer_ist_online_348x180 = 'Eisdiele_Wer_ist_online_348x180',
    Eisdiele_Neuigkeiten_348x180 = 'Eisdiele_Neuigkeiten_348x180',
    Eisdiele_Umfragen_348x180 = 'Eisdiele_Umfragen_348x180',
    Clubboard_Clubbesuch_500x500 = 'Clubboard_Clubbesuch_500x500',
    Clubboard_Cocktails_500x500 = 'Clubboard_Cocktails_500x500',
    Clubboard_Dinner_500x500 = 'Clubboard_Dinner_500x500',
    Clubboard_FotografieVideo_500x500 = 'Clubboard_FotografieVideo_500x500',
    Clubboard_Kino_500x500 = 'Clubboard_Kino_500x500',
    Clubboard_Paerchenabend_500x500 = 'Clubboard_Paerchenabend_500x500',
    Clubboard_Sauna_500x500 = 'Clubboard_Sauna_500x500',
    Clubboard_Sonstiges_500x500 = 'Clubboard_Sonstiges_500x500',
    Clubboard_Sport_500x500 = 'Clubboard_Sport_500x500',
    Clubboard_suche_FFM_500x500 = 'Clubboard_suche_FFM_500x500',
    Clubboard_suche_MFM_500x500 = 'Clubboard_suche_MFM_500x500',
    Clubboard_Swinger_500x500 = 'Clubboard_Swinger_500x500',
    Clubboard_Treffen_500x500 = 'Clubboard_Treffen_500x500',
    Clubboard_Urlaub_500x500 = 'Clubboard_Urlaub_500x500',
    Profilseiten_Defaulttitelbild1_all_1920x350 = 'Profilseiten_Defaulttitelbild1_all_1920x350',
    Profilseiten_Defaultsilhouettenbild1_Lady_320x320 = 'Profilseiten_Defaultsilhouettenbild1_Lady_320x320',
    Profilseiten_Defaultsilhouettenbild2_Lady_320x320 = 'Profilseiten_Defaultsilhouettenbild2_Lady_320x320',
    Profilseiten_Defaultsilhouettenbild3_Lady_320x320 = 'Profilseiten_Defaultsilhouettenbild3_Lady_320x320',
    Profilseiten_Defaultsilhouettenbild1_Gentleman_320x320 = 'Profilseiten_Defaultsilhouettenbild1_Gentleman_320x320',
    Profilseiten_Defaultsilhouettenbild2_Gentleman_320x320 = 'Profilseiten_Defaultsilhouettenbild2_Gentleman_320x320',
    Profilseiten_Defaultsilhouettenbild3_Gentleman_320x320 = 'Profilseiten_Defaultsilhouettenbild3_Gentleman_320x320',
    Profilseiten_Defaultsilhouettenbild1_Lovers_320x320 = 'Profilseiten_Defaultsilhouettenbild1_Lovers_320x320',
    Profilseiten_Defaultsilhouettenbild2_Lovers_320x320 = 'Profilseiten_Defaultsilhouettenbild2_Lovers_320x320',
    Profilseiten_Defaultsilhouettenbild3_Lovers_320x320 = 'Profilseiten_Defaultsilhouettenbild3_Lovers_320x320'

}